% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{busiest_routes}
\alias{busiest_routes}
\title{Compute busiest routes}
\usage{
busiest_routes(dataframe, origincol, destcol)
}
\arguments{
\item{dataframe}{data}

\item{origincol}{origin column}

\item{destcol}{destination column}
}
\value{
busiest routes
}
\description{
foo
}
\details{
foo
}
