% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load DB1B data}
\usage{
load_data(datafile, cityfile, carrierfile)
}
\arguments{
\item{datafile}{Data file name}

\item{cityfile}{City file name}

\item{carrierfile}{Carrier file name}
}
\value{
linked data
}
\description{
bar
}
\details{
bar
}
