% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{market_shares}
\alias{market_shares}
\title{Compute carrier market shares by airport}
\usage{
market_shares(dataframe, carriercol, origincol)
}
\description{
This function computes the market share of each carrier at each airport.
}
\details{
The carrier and airport columns can be specified manually, so ticketing or operating
carrier can be used (or another carrier column, if you had one---for instance, mainline
vs regional carrier market share, or legacy vs. low-cost). The airport column is likewise
an argument, so market shares per-airport, per-city, per-state, etc. can be calculated.
}
